refactor: split store into separate files + namespace store types
